fix: v1/util.tsの共通部分をcore/util.tsに移動
