+++ v1/test/match_test.js
@@ -1,8 +1,9 @@
 import { getAuth, signInWithEmailAndPassword } from "../../deps.ts";
 import { assert, assertEquals, v4 } from "../../deps-test.ts";
 
+import { randomUUID } from "../../core/util.ts";
+
 import ApiClient from "../../client/client.ts";
-import { randomUUID } from "../util.ts";
 
 const ac = new ApiClient();
