+++ v1/test/util_test.ts
@@ -1,5 +1,5 @@
 import { assert, v4 } from "../../deps-test.ts";
-import { randomUUID } from "../util.ts";
+import { randomUUID } from "../../core/util.ts";
 
 Deno.test("randamUUID", () => {
   const uuid = randomUUID();
